+++ tests/testthat/test-to-description.R
@@ -35,6 +35,23 @@ test_that("att_amend_desc updates description", {
 })
 unlink(dummypackage, recursive = TRUE)
 
+# Test Deprecated att_to_description ----
+# suppressWarnings()
+# Copy package in a temporary directory
+tmpdir <- tempdir()
+file.copy(system.file("dummypackage",package = "attachment"), tmpdir, recursive = TRUE)
+dummypackage <- file.path(tmpdir, "dummypackage")
+# browseURL(dummypackage)
+suppressWarnings(att_to_description(path = dummypackage))
+desc_file <- readLines(file.path(tmpdir, "dummypackage", "DESCRIPTION"))
+namespace_file <- readLines(file.path(tmpdir, "dummypackage", "NAMESPACE"))
+
+test_that("att_to_description still updates namespace", {
+  expect_length(namespace_file, 4)
+})
+unlink(dummypackage, recursive = TRUE)
+
+
 # att_to_desc_from_is ----
 # Copy package in a temporary directory
 tmpdir <- tempdir()
@@ -42,7 +59,7 @@ file.copy(system.file("dummypackage",package = "attachment"), tmpdir, recursive 
 dummypackage <- file.path(tmpdir, "dummypackage")
 # browseURL(dummypackage)
 att_to_desc_from_is(path.d = file.path(dummypackage, "DESCRIPTION"),
-                      imports = c("magrittr", "attachment"), suggests = c("knitr"))
+                    imports = c("magrittr", "attachment"), suggests = c("knitr"))
 
 desc_file <- readLines(file.path(tmpdir, "dummypackage", "DESCRIPTION"))
 
@@ -62,153 +79,7 @@ test_that("att_to_desc_from_is updates description", {
 })
 unlink(dummypackage, recursive = TRUE)
 
-# Test Deprecated ----
-# suppressWarnings()
-# Copy package in a temporary directory
-tmpdir <- tempdir()
-file.copy(system.file("dummypackage",package = "attachment"), tmpdir, recursive = TRUE)
-dummypackage <- file.path(tmpdir, "dummypackage")
-# browseURL(dummypackage)
-suppressWarnings(att_to_description(path = dummypackage))
-desc_file <- readLines(file.path(tmpdir, "dummypackage", "DESCRIPTION"))
-namespace_file <- readLines(file.path(tmpdir, "dummypackage", "NAMESPACE"))
 
-test_that("att_to_description still updates namespace", {
-  expect_length(namespace_file, 4)
-})
-unlink(dummypackage, recursive = TRUE)
-
-
-# Test extract Remotes ----
-test_that("find_remotes works with no error", {
-  # Cannot verify directly if this works because it depends on user installation
-  # From CRAN or Github. At least there should be no error.
-  expect_true(length(find_remotes("desc")) == 0 |
-                          length(find_remotes("desc") == 1))
-  # base package avoided
-  expect_true(length(find_remotes("stats")) == 0)
-})
-
-# Test core of find_remotes ----
-test_that("extract_pkg_info extracts code", {
-  # Github
-  fake_desc_github <- list(
-    list(
-      RemoteType = "github",
-      RemoteHost = "api.github.com",
-      RemoteRepo = "golem",
-      RemoteUsername = "ThinkR-open"
-    )
-  ) %>% setNames("golem")
-  expect_equal(extract_pkg_info(fake_desc_github)[["golem"]], "thinkr-open/golem")
-
-  # GitLab
-  # Sys.setenv(GITLAB_PAT = "xxxxxxxxxxxxxxxx")
-  # remotes::gitlab_pat(FALSE)
-  # remotes::install_gitlab("statnmap/fakepkg")
-  fake_desc_gitlab <- list(
-    list(
-      RemoteType = "gitlab",
-      RemoteHost = "gitlab.com",
-      RemoteRepo = "fakepkg",
-      RemoteUsername = "statnmap"
-    )
-  ) %>% setNames("fakepkg")
-  expect_equal(extract_pkg_info(fake_desc_gitlab)[["fakepkg"]], "gitlab::statnmap/fakepkg")
-
-  # Other installations
-  fake_desc_local <- list(
-    list(
-      RemoteType = NULL,
-      RemoteHost = NULL,
-      RemoteRepo = NULL,
-      RemoteUsername = NULL
-    )
-  ) %>% setNames("fakenull")
-
-  expect_true(is.na(extract_pkg_info(fake_desc_local)[["fakenull"]]))
-  expect_equal(names(extract_pkg_info(fake_desc_local)[["fakenull"]]), "local maybe ?")
-
-  # Test internal_remotes_to_desc ----
-  tmpdir <- tempdir()
-  file.copy(system.file("dummypackage",package = "attachment"), tmpdir, recursive = TRUE)
-  dummypackage <- file.path(tmpdir, "dummypackage")
-
-  path.d <- file.path(dummypackage, "DESCRIPTION")
-  cat("Remotes:\n    thinkr-open/attachment", append = TRUE,
-      file = path.d)
-
-  remotes <- c(
-    extract_pkg_info(fake_desc_github),
-    extract_pkg_info(fake_desc_gitlab),
-    extract_pkg_info(fake_desc_local)
-  )
-
-  expect_error(
-    internal_remotes_to_desc(remotes, path.d = path.d, stop_local = TRUE),
-    "installed from source locally"
-  )
-
-  expect_message(
-    expect_message(
-      internal_remotes_to_desc(remotes, path.d = path.d, stop_local = FALSE),
-      "installed from source locally"
-    ),
-    "Remotes for attachment, golem & fakepkg were added to DESCRIPTION."
-  )
-
-  new_desc <- readLines(path.d)
-
-  w.remotes <- grep('Remotes:', new_desc)
-  expect_length(w.remotes, 1)
-  expect_equal(new_desc[w.remotes + 1], "    thinkr-open/attachment,")
-  expect_equal(new_desc[w.remotes + 2], "    thinkr-open/golem,")
-  expect_equal(new_desc[w.remotes + 3], "    gitlab::statnmap/fakepkg")
-
-  unlink(dummypackage, recursive = TRUE)
-})
-
-# test set_remotes_to_desc ----
-tmpdir <- tempdir()
-file.copy(system.file("dummypackage",package = "attachment"), tmpdir, recursive = TRUE)
-dummypackage <- file.path(tmpdir, "dummypackage")
-
-test_that("set_remotes_to_desc return nothing if local installs", {
-
-  skip_on_cran()
-  # We do not know whether some packages are installed manually from source on CRAN
-
-    pkgs <- att_amend_desc(dummypackage) %>%
-      att_from_description()
-    remotes <- find_remotes(pkgs)
-
-    desc_file <- file.path(dummypackage, "DESCRIPTION")
-
-    if (is.null(remotes)) {
-      # Do not test if some are compiled locally
-      expect_message(
-        set_remotes_to_desc(desc_file),
-        "no remote packages installed"
-      )
-    } else {
-      pkgnames <- glue::glue_collapse(names(remotes), sep = ", ", last = " & ")
-      nona <- unlist(lapply(remotes, is.na))
-      expect_message(
-        set_remotes_to_desc(desc_file),
-        paste("Remotes for", pkgnames[nona])
-      )
-    }
-
-    # Add attachment in DESCRIPTION as should be local installed
-    desc_lines <- readLines(desc_file)
-    desc_lines[desc_lines == "Suggests: "] <- "Suggests: \n    attachment,"
-    writeLines(desc_lines, con = desc_file)
-
-    expect_message(
-      set_remotes_to_desc(desc_file),
-      "Package attachment was probably installed from source locally"
-    )
-})
 
 # Test missing DESCRIPTION works ----
 # Copy package in a temporary directory
@@ -256,3 +127,137 @@ test_that("Works with missing DESCRIPTION", {
   expect_true(file.exists(file.path(dummypackage, "NAMESPACE")))
 
 })
+
+# Note: glue is necessary for {attachment}, ggplot2 is not even in suggests.
+
+# missing pkg in R is not installed ----
+test_that("missing pkg is not installed", {
+  # Copy package in a temporary directory
+  tmpdir <- tempfile("dummy")
+  dir.create(tmpdir)
+  file.copy(system.file("dummypackage",package = "attachment"), tmpdir, recursive = TRUE)
+  dummypackage <- file.path(tmpdir, "dummypackage")
+  # browseURL(dummypackage)
+
+  cat("
+#' Function
+#' @importFrom ggplot ggplot
+#' @export
+my_fun <- function() {
+data %>%
+filter(col == 3) %>%
+mutate(new_col = 1) %>%
+ggplot() +
+  aes(x, y, colour = new_col) +
+  geom_point()
+}
+", file = file.path(dummypackage, "R", "function.R"))
+
+  expect_error(attachment::att_amend_desc(path = dummypackage),
+               "The package ggplot is missing or misspelled.")
+
+  # Clean after
+  unlink(dummypackage)
+})
+
+# missing pkgS are not installed ----
+test_that("missing pkgS are not installed", {
+
+  # Copy package in a temporary directory
+  tmpdir <- tempfile("dummy")
+  dir.create(tmpdir)
+  file.copy(system.file("dummypackage",package = "attachment"), tmpdir, recursive = TRUE)
+  dummypackage <- file.path(tmpdir, "dummypackage")
+
+  cat("
+#' Function
+#' @importFrom ggplot ggplot
+#' @importFrom ggplot3 ggplot
+# @importFrom ggplot4 ggplot # Only comment, not used
+#' @export
+my_fun <- function() {
+data %>%
+filter(col == 3) %>%
+mutate(new_col = 1) %>%
+ggplot() +
+  aes(x, y, colour = new_col) +
+  geom_point()
+}
+", file = file.path(dummypackage, "R", "function.R"))
+
+  expect_error(attachment::att_amend_desc(path = dummypackage),
+               "Packages ggplot & ggplot3 are missing or misspelled.")
+
+  # Clean after
+  unlink(dummypackage)
+})
+
+# missing pkg in vignette (Suggests) is not installed ----
+test_that("missing pkg is not installed", {
+  # Copy package in a temporary directory
+  tmpdir <- tempfile("dummy")
+  dir.create(tmpdir)
+  file.copy(system.file("dummypackage",package = "attachment"), tmpdir, recursive = TRUE)
+  dummypackage <- file.path(tmpdir, "dummypackage")
+  # browseURL(dummypackage)
+
+  cat("
+## The vignette
+```{r}
+library(glue)
+library(ggplot3)
+```
+", file = file.path(dummypackage, "vignettes", "vignette.Rmd"))
+
+  expect_error(attachment::att_amend_desc(path = dummypackage),
+               "The package ggplot3 is missing or misspelled.")
+
+  # Clean after
+  unlink(dummypackage)
+})
+
+# Test update desc when missing packages for books ----
+tmpdir <- tempfile("miss")
+dir.create(tmpdir)
+file.copy(system.file("dummypackage",package = "attachment"),
+          tmpdir, recursive = TRUE)
+dummypackage <- file.path(tmpdir, "dummypackage")
+
+test_that("att_to_desc_from_is can update DESCRIPTION w/o uninstalled packages", {
+  imports <- unique(
+    c("fakeinstalled", # Not exists
+      "bookdown", "knitr",
+      "pagedown", # Not installed on CI usually
+      att_from_rmds(file.path(dummypackage, "vignettes"), recursive = FALSE))
+  )
+  # Default to error
+  expect_error(att_to_desc_from_is(
+    path.d = file.path(dummypackage, "DESCRIPTION"),
+    imports = imports
+  ), regexp = "missing or misspelled")
+
+  # must.exist = FALSE
+  expect_warning(
+    att_to_desc_from_is(
+      path.d = file.path(dummypackage, "DESCRIPTION"),
+      imports = imports,
+      must.exist = NA # then warning
+    ),
+    regexp = "missing or misspelled")
+
+  expect_warning(
+    expect_error(
+      att_to_desc_from_is(
+        path.d = file.path(dummypackage, "DESCRIPTION"),
+        imports = imports,
+        must.exist = FALSE # then no error, no warning
+      ), regexp = NA),
+    regexp = NA)
+
+  desc <- readLines(file.path(file.path(dummypackage, "DESCRIPTION")))
+  expect_true(any(grepl("fakeinstalled", desc)))
+  expect_true(any(grepl("bookdown", desc)))
+  expect_true(any(grepl("knitr", desc)))
+  expect_true(any(grepl("pagedown", desc)))
+})
+unlink(tmpdir, recursive = TRUE)